Port to Python 3 and Django 2.1 (#5)
* Ignore build directory in git

* Install django_cubrid for python >= 3

* Fix some comparisons with None

* Advance django_cubrid version from 0.0.1 to 0.1.0

* Adapt print statements for python 3

* Adapt exceptions for python 3

* Remove leftover data_types dictionary

It was moved in base.py. This version is redundant and obsoleted and not
used anywhere anymore.

* Port sys.maxint to sys.maxsize for python 3

* Place context methods for the cursors for newer Django versions

* Add optional name argument for create_cursor

Necessary for newer Django versions. In our case, if the arg is given,
it is ignored.

* Adapt DatabaseWrapper's component creation for Django >= 1.11

For these newer versions, the types for the components are defined in
the class using special attributes.

* Adapt bulk_insert_sql for Django >= 1.9

* Django Field.rel attribute is now remote_field

To test if it is a relationship field, the is_relation attribute can be
used.

* Implement introspection get_constraints method (Django >= 1.7)

Sometimes it is needed, otherwise errors occur in the migrations.

* Fix binary field - does not allow null/unique

* Adapt boolean fields to small integer values

* Implement case-insensitive lookups using UPPER

The iLIKE operator is not supported by CUBRID, so using UPPER is an
alternative, taken from the PosgreSQL driver.