Fix type conversion issues with some Django field types
Some field types need conversions performed in order to behave as
expected with Django. Including BooleanField, NullBooleanField,
TextField, DateTimeFields, UUIDFields.

There was an older approach to converting the date-time values, but it
is simpler with the new approach.

The new approach is copied from the MySql django backend, which needs
the same conversions.